feat: refactor CLI logic out to it's own package under core/cli
